+++ packages/client-core/src/components/Settings/index.tsx
@@ -43,11 +43,11 @@ import WorldSettings from './WorldSettings'
 import { useNavigationProvider } from '../Glass/NavigationProvider'
 import AvatarScreen from './AvatarScreen'
 import DeleteAccountScreen from './DeleteAccountScreen'
-import DisplayNameScreen from './DisplayNameScreen'
 import PermissionsScreen from './PermissionsScreen'
 import ShareSpaceScreen from './ShareSpaceScreen'
 import SignUpScreen from './SignUpScreen'
 import SSOScreen from './SSOScreen'
+import UsernameScreen from './UsernameScreen'
 
 // Define screen structure type
 interface ScreenDefinition {
@@ -82,8 +82,8 @@ export const screens: Record<string, ScreenDefinition> = {
     component: (props) => <PlaceholderScreen {...props} title="Controls" />,
     title: 'Controls'
   },
-  displayName: {
-    component: DisplayNameScreen,
+  username: {
+    component: UsernameScreen,
     title: 'Display Name'
   },
   userId: {
